move camera manipulation stuff out to a new utils module
